Move Hierarchy to hierarchy.js
